backend/requirements.txt
# Core FastAPI & Server
fastapi==0.115.12
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (PostgreSQL 17 지원)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client for External Auth & API Provider
httpx==0.28.1

# File Handling & Static Files
aiofiles==23.2.1

# Scheduler (MVP 모듈 관리용)
apscheduler==3.10.4

# System Monitoring
psutil==5.9.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
python-dotenv==1.0.0

# Logging
structlog==23.1.0

# MSSQL Support (for multi-database data providers)
aioodbc==0.5.0
pyodbc==5.2.0